Task service implementation
