+++ README.md
@@ -1,3 +1,9 @@
 # Alamo
 
-This project was developed with Angular v18.
+This project was developed with Angular v18. I am using Material UI as default library, Ionicons is also used for special icons like 
+the GitHub icon in the top right corner of the page.
+
+## Folder structure
+
+The structure of this project is based on features with a feature folder group all available features. Then we have a shared folder 
+which is responsible for components in common like toolbar or footer.
